Moving handlers to new file
